add optional updateMainThread callback support
Once `init` returns (and it must return, as not all platforms
allow `main` to block), then Mach calls `update` on a secondary
OS thread which is used for rendering - while e.g. the main
thread remains focused on input handling. This enables graphics
rendering and input to operate at different frequencies, which
is very important for being able to handle input events mid-frame
or at a higher frequency than rendering in general.

Still, there are cases where one really must execute code on the
main thread - such as if you are making a desktop GUI app and
wish to open a native file dialog.

To address this need, `pub fn updateMainThread` is now an optional
function your app can declare. It has the same signature as
`pub fn update`, the only difference being it is invoked on the
main OS thread.

If there is no input/rendering thread separation, or no main OS
thread is accessible on the platform, such as in browsers and
potentially on mobile devices - then `updateMainThread` will be
called in sequence with `update` and from the same thread as
`update`.

This is an optional callback, and in general its use should be
avoided as it must be multiplexed with input handling (what you
do in `updateMainThread` blocks input handling, vice-versa.)

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>