switch back to `App.init` taking a `*App` to initialize
In the API redesign, we changed this:

```zig
pub fn init(app: *App) !void {
    app.* = App{
        // ...
    };
}
```

To this:

```zig
pub fn init() !App {
    return App{
        // ...
    };
}
```

On the surface, that looks nicer and more appealing. However, it eliminates
a common use case which is desirable: using `App` as backing-storage and taking
pointers to it. For example:

```
pub const App = @This();

color_attachments: [2]gpu.RenderPassColorAttachment,
descriptor: gpu.RenderPassDescriptor,

pub fn init() !App {
    var app: App = undefined;
    app.color_attachments = [_]gpu.RenderPassColorAttachment{...};
    app.descriptor = gpu.RenderPassDescriptor.init(.{
        .color_attachments = &app.color_attachments, // pointer to `app` memory
    });
    return app; // now has a dangling reference to stack memory
}
```

As a result, we undo the change and go back to a `pub fn init(app: *App) !void`
model in which you must initialize the undefined memory `app` which is passed
into the function.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>