examples: Make triangle and MSAA consistent.
Previously, there were several minor differences between the "triangle"
and "triangle-msaa" examples which distracted from the core changes
between the two. Now, the difference between the two represents the user
journey from the more simple example to the more complex one.

Additionally, fix two pedantic issues with resource management. There
was a theoretical resource leak in the event of an error after creating
the pipeline, and the release calls in `App.deinit` were ordered in FIFO
order when they should probably be LIFO.